chore(preset-wind): organize blend-modes (#401)
