+++ packages/preset-wind/src/rules/background.ts
@@ -40,24 +40,6 @@ const bgGradientColorResolver = (mode: 'from' | 'to' | 'via') =>
   }
 
 export const backgroundStyles: Rule[] = [
-  // blends
-  ['bg-blend-multiply', { 'background-blend-mode': 'multiply' }],
-  ['bg-blend-screen', { 'background-blend-mode': 'screen' }],
-  ['bg-blend-overlay', { 'background-blend-mode': 'overlay' }],
-  ['bg-blend-darken', { 'background-blend-mode': 'darken' }],
-  ['bg-blend-lighten', { 'background-blend-mode': 'lighten' }],
-  ['bg-blend-color-dodge', { 'background-blend-mode': 'color-dodge' }],
-  ['bg-blend-color-burn', { 'background-blend-mode': 'color-burn' }],
-  ['bg-blend-hard-light', { 'background-blend-mode': 'hard-light' }],
-  ['bg-blend-soft-light', { 'background-blend-mode': 'soft-light' }],
-  ['bg-blend-difference', { 'background-blend-mode': 'difference' }],
-  ['bg-blend-exclusion', { 'background-blend-mode': 'exclusion' }],
-  ['bg-blend-hue', { 'background-blend-mode': 'hue' }],
-  ['bg-blend-saturation', { 'background-blend-mode': 'saturation' }],
-  ['bg-blend-color', { 'background-blend-mode': 'color' }],
-  ['bg-blend-luminosity', { 'background-blend-mode': 'luminosity' }],
-  ['bg-blend-normal', { 'background-blend-mode': 'normal' }],
-
   // TODO: bg-[url] { background-image: x }
 
   // gradients
