+++ packages/preset-wind/src/rules/index.ts
@@ -9,7 +9,6 @@ export * from './divide'
 export * from './filters'
 export * from './line-clamp'
 export * from './scrolls'
-export * from './shadow'
 export * from './spacing'
 export * from './static'
 export * from './table'
